requirements.txt
pytorch
matplotlib
pandas
numpy
torchvision
torchaudio
scikit-learn
torchmetrics
tqdm
mlxtend
torchinfo
gradio